requirements.txt
asgiref==3.5.2
Django==4.0.8
psycopg2-binary==2.9.3
sqlparse==0.4.3
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
django-allauth==0.50.0
environs[django]==9.5.0
pillow==9.3.0
